core: remove math expression, add include/using and fn calls.
Introducing a new thing called "keywords"... yeah I'm terrible
at naming. Basically commands to tell MDiocre something else than
getting and setting variables.

The include function is something that MDiocre needed, but for
now the implementation is still a bit wonky and depends on which
directory (pwd) is active when the user calls the function.

There is also the "using" call for when you want to define some
functions, which you can use with the new parentheses operand in
the assignment statement, ex: (toRFC822 pubdate) would be equivalent
to calling toRFC822(pubdate) in regular Python.

That said the latter is probably something to be exploited. I'm
terrible lmao.